discard/requirements.txt
gunicorn==20.1.0
Flask==2.2.5
flask-socketio==5.3.6
python-socketio==5.11.1
python-engineio==4.8.2
pymongo==4.6.1
opencv-python==4.9.0.80
mediapipe==0.10.9
numpy==1.26.4
flask-cors
